requirements.txt
black==22.3.0
flake8==4.0.1
wheel==0.40.0
twine==4.0.2
pytube==15.0.0
fastapi==0.101.0
uvicorn==0.23.2
# https://download.pytorch.org/whl/cu101/torch-1.6.0+cu101-cp38-cp38-linux_x86_64.whl #+wrecking the url
# https://download.pytorch.org/whl/cu101/torch-1.6.0%2Bcu101-cp38-cp38-linux_x86_64.whl #still sort of gpu be even better
# https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp38-cp38-linux_x86_64.whl # Has weird error https://github.com/NVIDIA/NeMo/issues/1590
#torch @ https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp38-cp38-linux_x86_64.whl; platform_system=="Linux"
#torch @ https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp38-cp38-win_amd64.whl; platform_system=="Windows"
openai==0.27.7
openai-whisper==20230314